// src/services/types/content-folder-node-response.ts
/** biome-ignore-all lint/correctness/noUnusedImports: IDONCARE */
/** biome-ignore-all lint/suspicious/noExplicitAny: IDONCARE */
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 * Source: api-docs
 * Title: CMS API
 * Description: CMS API
 * OpenAPI spec version: v1.0.0
 */

export const contentFolderNodeResponseNodeTypeEnum = {FOLDER: "FOLDER", TABLE: "TABLE"} as const;

export type ContentFolderNodeResponseNodeTypeEnumKey =
	(typeof contentFolderNodeResponseNodeTypeEnum)[keyof typeof contentFolderNodeResponseNodeTypeEnum];

export type ContentFolderNodeResponse = {
	/**
	 * @description 노드 아이디
	 * @type integer, int64
	 */
	id: number;
	/**
	 * @description 부모 폴더 아이디
	 * @type integer | undefined, int64
	 */
	parentId?: number;
	/**
	 * @description 노드 타입(FOLDER/TABLE)
	 * @type string
	 */
	nodeType: ContentFolderNodeResponseNodeTypeEnumKey;
	/**
	 * @description 표시명
	 * @type string
	 */
	name: string;
	/**
	 * @description 테이블명(TABLE 타입 전용)
	 * @type string | undefined
	 */
	tableName?: string;
	/**
	 * @description 정렬 순서
	 * @type integer, int32
	 */
	orderNumber: number;
	/**
	 * @description 노출 여부
	 * @type boolean
	 */
	displayed: boolean;
	/**
	 * @description 하위 노드
	 * @type object
	 */
	children: ContentFolderNodeResponse;
};