// src/services/types/content-folder-save-node-request.ts
/** biome-ignore-all lint/correctness/noUnusedImports: IDONCARE */
/** biome-ignore-all lint/suspicious/noExplicitAny: IDONCARE */
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 * Source: api-docs
 * Title: CMS API
 * Description: CMS API
 * OpenAPI spec version: v1.0.0
 */

export const contentFolderSaveNodeRequestNodeTypeEnum = {FOLDER: "FOLDER", TABLE: "TABLE"} as const;

export type ContentFolderSaveNodeRequestNodeTypeEnumKey =
	(typeof contentFolderSaveNodeRequestNodeTypeEnum)[keyof typeof contentFolderSaveNodeRequestNodeTypeEnum];

export type ContentFolderSaveNodeRequest = {
	/**
	 * @description 노드 타입(FOLDER/TABLE)
	 * @type string
	 */
	nodeType: ContentFolderSaveNodeRequestNodeTypeEnumKey;
	/**
	 * @description 표시명(FOLDER/TABLE에서 필수)
	 * @minLength 1
	 * @type string
	 */
	name: string;
	/**
	 * @description 테이블명(TABLE에서 필수)
	 * @type string | undefined
	 */
	tableName?: string;
	/**
	 * @description 노출 여부
	 * @type boolean
	 */
	displayed: boolean;
	/**
	 * @description 하위 노드(FOLDER에서만 사용, TABLE만 허용)
	 * @type array | undefined
	 */
	children?: ContentFolderSaveNodeRequest[];
};